HR Agent/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
chromadb
streamlit
python-dotenv
langfuse
